AuthCodes Generator and Directory Remanagement
I've finished the creating of random AuthCodes and the Process of Checking them on the Login/Registration Page.

The email functionality is not present on the uni's webserver so unfortunately, the user will just have to copy and paste their AuthCode to the chosen person. (Maybe have a list of AuthCodes generated from that user present to them in the future?)

Also, I have done a bit of rejigging of the files and directories of the whole repo. Now the navbar.php and .js file is at the 'root' and simply grabs the html/php files from the individual folders when it needs to display them (building from Charlie's idea of not having separate pages but instead it all revolving around the navbar).

Let me know of anything you're confused about/ think needs improvement